requirements.txt
# Core FastAPI dependencies
fastapi>=0.128.0
uvicorn>=0.40.0
python-multipart>=0.0.9
itsdangerous>=2.1.2

# Database
sqlalchemy>=2.0.46
psycopg2-binary>=2.9.9
alembic>=1.18.1

# Environment & Configuration
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic[email]>=2.12.5
pydantic-settings>=2.12.0

# Document Storage (S3/MinIO)
boto3>=1.42.36

# LangChain & AI
langchain>=1.2.7
langchain-openai>=1.1.7
langchain-core>=1.2.7
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=1.0.7

# OpenAI
openai>=2.15.0
tiktoken>=0.12.0

# MCP (Model Context Protocol)
mcp>=1.26.0
fastmcp>=2.14.4

# Utilities
httpx>=0.28.1
requests>=2.32.5
pydantic-core>=2.41.5
typing-extensions>=4.15.0
typing-inspection>=0.4.2
email-validator>=2.3.0

# Authentication & Security
pyjwt>=2.10.1
cryptography>=46.0.4
argon2-cffi>=23.1.0
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Email
aiosmtplib>=3.0.1
jinja2>=3.1.3
sendgrid==6.11.0

# PDF Generation
weasyprint>=60.0
zipstream-ng>=1.7.0

# Rate Limiting
slowapi>=0.1.9

# User Agent Parsing (for session tracking)
user-agents>=2.2.0

# Monitoring & Observability
opentelemetry-api>=1.39.1
opentelemetry-sdk>=1.39.1
opentelemetry-instrumentation>=0.60b1
prometheus-client>=0.24.1

# Data Processing
jsonschema>=4.26.0
pyyaml>=6.0.3
ormsgpack>=1.12.1
orjson>=3.11.5

# Development
rich>=14.3.1

# Windows-specific (optional)
pywin32>=311; sys_platform == 'win32'